Membuat emptyObjek dan memperbaiki case jumlah input stok saat peminjaman tidak sesuai

+++ peminjaman_alat_app/lib/features/alat/widgets/inputStokAjukan_button.dart
@@ -50,7 +50,7 @@ class _inputStokDanButtonAjukanState extends State<inputStokDanButtonAjukan> {
           maxVal: widget.widget.stok,
           onQtyChanged: (valueStok) {
             setState(() {
-              jumlah = valueStok;
+              jumlah = valueStok.toInt();
             });
           },
         ),
